requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
bcrypt
pyjwt
pydantic[email]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart